제품 기능 추가

+++ src/main/java/com/hhplush/eCommerce/api/products/dto/response/ResponseProductListDTO.java
@@ -1,13 +1,13 @@
 package com.hhplush.eCommerce.api.products.dto.response;
 
-import com.hhplush.eCommerce.domain.enums.ProductState;
+import com.hhplush.eCommerce.domain.product.ProductState;
 import lombok.Builder;
 
 @Builder
 public record ResponseProductListDTO(
-    Integer productId,
+    Long productId,
     String productName,
-    Integer price,
+    Long price,
     ProductState productState
 ) {
